src/retrieval-api/requirements.txt
flask
gunicorn
google-cloud-firestore
google-cloud-aiplatform
langchain==0.1.0
langchain-community==0.0.10
langchain-google-vertexai